refactor for multi-host support
With a router plus home server, it is common to want to address multiple hosts
by their IPv6 address directly.
Adapt the units to allow to specify which host to forward to, and the resolver
to resolve multiple hosts.
Also add a description for each host which can be used in notifications,
and add some alternative solutions to the README.